pyproject.toml
[project]
name = "chat-bot-ui-v2"
version = "0.1.0"
description = ""
authors = [
    {name = "Yash Kushwaha",email = "37941798+YashKushwaha@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "uvicorn (>=0.34.3,<0.35.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "llama-index-llms-ollama (>=0.6.2,<0.7.0)",
    "notebook (>=7.4.3,<8.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "transformers[torch] (>=4.52.4,<5.0.0)",
    "llama-index-utils-workflow (>=0.3.5,<0.4.0)",
    "mlflow (>=3.1.0,<4.0.0)",
    "llama-index-llms-openai (>=0.4.7,<0.5.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "boto3 (==1.37.3)",
    "llama-index-llms-bedrock (>=0.3.8,<0.4.0)",
    "llama-index-llms-bedrock-converse (>=0.7.2,<0.8.0)",
    "llama-index-embeddings-huggingface (>=0.5.4,<0.6.0)",
    "datasets (>=3.6.0,<4.0.0)",
    "llama-index-vector-stores-chroma (>=0.4.2,<0.5.0)",
    "protobuf (==3.20.3)",
    "ipywidgets (>=8.1.7,<9.0.0)",
    "pymongo (>=4.13.2,<5.0.0)"
]
[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
